some changes in migrations

+++ db/migrate/20101203101417_create_reservations.rb
@@ -1,16 +1,18 @@
 class CreateReservations < ActiveRecord::Migration
   def self.up
-     create_table :reservations do |t|
-       t.integer :restaurant_id
-       t.references :user
-       t.integer :status_reservation_id
-       t.timestamps
-       t.datetime :date
-     end
+    unless ActiveRecord::Base.connection.table_exists? :reservations
+      create_table :reservations do |t|
+        t.integer :restaurant_id
+        t.references :user
+        t.integer :status_reservation_id
+        t.timestamps
+        t.datetime :date
+      end
+    end
   end
 
   def self.down
-    remove_column :reservations
+    remove_column :reservations if ActiveRecord::Base.connection.table_exists? :reservations
   end
 end
